# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestión de Productos API",
        "description": "API para gestionar productos y sus precios",
        "contact": {
            "email": "soporte@tudominio.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/currencies": {
            "get": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Obtener todas las monedas",
                "description": "Devuelve una lista de todas las monedas.",
                "operationId": "73e0a13248f9f6aea618716227eab68a",
                "responses": {
                    "200": {
                        "description": "Lista de monedas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Currency"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Crear una nueva moneda",
                "description": "Crea una nueva moneda y la devuelve.",
                "operationId": "57e1c3f26ae9c82bd5996ac5bb738c5b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Currency"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Moneda creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Currency"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/currencies/{id}": {
            "get": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Obtener una moneda por ID",
                "description": "Devuelve los detalles de una moneda específica.",
                "operationId": "23440d4caad0c601f680b655f9c907f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la moneda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la moneda",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Currency"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Moneda no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Actualizar una moneda existente",
                "description": "Actualiza los detalles de una moneda existente.",
                "operationId": "8a397517013e49ac5bf2db204caaa130",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la moneda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Currency"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Moneda actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Currency"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Moneda no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Eliminar una moneda",
                "description": "Elimina una moneda específica por su ID.",
                "operationId": "c5d45f11f33d23039acf71db15a4cf31",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la moneda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Moneda eliminada exitosamente"
                    },
                    "404": {
                        "description": "Moneda no encontrada"
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener lista de productos",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "responses": {
                    "200": {
                        "description": "Lista de productos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un nuevo producto",
                "operationId": "e8185e307706e2cd84bbcf5171e2f09d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener un producto por ID",
                "operationId": "7ab21ddb713501f3d4c2914892c02a21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del producto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar un producto por ID",
                "operationId": "6a7e11900844cd85cff66e662b82bd4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar un producto por ID",
                "operationId": "e80a6ca46716acee6b47ea1ac91663c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Producto eliminado exitosamente"
                    }
                }
            }
        },
        "/api/products/{productId}/prices": {
            "get": {
                "tags": [
                    "Product Prices"
                ],
                "summary": "Obtener los precios de un producto",
                "description": "Devuelve una lista de los precios asociados a un producto.",
                "operationId": "9795682918251db3525a68361862426f",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de precios del producto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductPrice"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Product Prices"
                ],
                "summary": "Crear un nuevo precio para un producto",
                "description": "Crea un nuevo precio asociado a un producto.",
                "operationId": "e607c4b0afa1b3f2dfbb98a8dda32834",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductPrice"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Precio creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductPrice"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Currency": {
                "required": [
                    "name",
                    "symbol",
                    "exchange_rate"
                ],
                "properties": {
                    "id": {
                        "description": "ID único de la moneda",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nombre de la moneda",
                        "type": "string"
                    },
                    "symbol": {
                        "description": "Símbolo de la moneda",
                        "type": "string"
                    },
                    "exchange_rate": {
                        "description": "Tipo de cambio de la moneda",
                        "type": "number",
                        "format": "float"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Product": {
                "required": [
                    "name",
                    "description",
                    "price",
                    "currency_id",
                    "tax_cost",
                    "manufacturing_cost"
                ],
                "properties": {
                    "id": {
                        "description": "Identificador único del producto",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nombre del producto",
                        "type": "string"
                    },
                    "description": {
                        "description": "Descripción del producto",
                        "type": "string"
                    },
                    "price": {
                        "description": "Precio del producto en la divisa base",
                        "type": "number",
                        "format": "float"
                    },
                    "currency_id": {
                        "description": "ID de la divisa base",
                        "type": "integer"
                    },
                    "tax_cost": {
                        "description": "Costo de impuestos del producto",
                        "type": "number",
                        "format": "float"
                    },
                    "manufacturing_cost": {
                        "description": "Costo de fabricación del producto",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
            "ProductPrice": {
                "required": [
                    "price",
                    "currency_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID único del precio",
                        "type": "integer"
                    },
                    "price": {
                        "description": "Precio del producto",
                        "type": "number",
                        "format": "float"
                    },
                    "currency_id": {
                        "description": "ID de la divisa asociada",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Productos",
            "description": "APIs relacionadas con la gestión de productos"
        },
        {
            "name": "Currencies",
            "description": "Currencies"
        },
        {
            "name": "Product Prices",
            "description": "Product Prices"
        }
    ]
}